feat(tree-grid): fixing some groupBy related classes #2530
