+++ projects/igniteui-angular/src/lib/grids/grid.common.ts
@@ -561,51 +561,6 @@ export class IgxColumnMovingDropDirective extends IgxDropDirective implements On
         }
     }
 }
-/**
- * @hidden
- */
-@Directive({
-    selector: '[igxGroupAreaDrop]'
-})
-export class IgxGroupAreaDropDirective extends IgxDropDirective {
-
-    constructor(private elementRef: ElementRef, private renderer: Renderer2, private zone: NgZone) {
-        super(elementRef, renderer, zone);
-    }
-
-    @HostBinding('class.igx-drop-area--hover')
-    public hovered = false;
-
-
-    public onDragEnter(event) {
-        const drag: IgxColumnMovingDragDirective = event.detail.owner;
-        const column: IgxColumnComponent = drag.column;
-        const isGrouped = column.grid.groupingExpressions.findIndex((item) => item.fieldName === column.field) !== -1;
-        if (column.groupable && !isGrouped) {
-            drag.icon.innerText = 'group_work';
-            this.hovered = true;
-        } else {
-            drag.icon.innerText = 'block';
-            this.hovered = false;
-        }
-    }
-
-    public onDragLeave(event) {
-        event.detail.owner.icon.innerText = 'block';
-        this.hovered = false;
-    }
-
-    public onDragDrop(event) {
-        const drag: IgxColumnMovingDragDirective = event.detail.owner;
-        if (drag instanceof IgxColumnMovingDragDirective) {
-            const column: IgxColumnComponent = drag.column;
-            const isGrouped = column.grid.groupingExpressions.findIndex((item) => item.fieldName === column.field) !== -1;
-            if (column.groupable && !isGrouped) {
-                column.grid.groupBy({ fieldName: column.field, dir: SortingDirection.Asc, ignoreCase: column.sortingIgnoreCase });
-            }
-        }
-    }
-}
 
 /**
  *@hidden
